bug: add a list of Bug class
The XML-RPC API for Bugzilla allows the use of several bug numbers for
some request such as 'get' and 'update'.
When dealing with a log of bugs, using this class will strongly improve
performance as less request will actually be sent and response received.

Only basic methods are implemented for now.

IMPORTANT NOTES: The API only allows to update all the given bugs with the
same value. This must stay in mind when using this new class.

Signed-off-by: Thomas Faivre <thomas.faivre@6wind.com>